src/backend/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-aiplatform>=1.38.0
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0